# open-api/passtheo-content-service-openapi.yaml
openapi: 3.1.0
info:
  title: PassTheo Content Service
  description: |
    The core learning engine for PassTheo. Serves questions, tracks progress,
    runs mock exams, manages streaks, awards achievements, and generates study plans.
    
    Content hierarchy: Country → ProductType → Product → Domain → Topic → Question
    Language is Strapi i18n, orthogonal to hierarchy.
  version: 1.0.0
servers:
  - url: http://localhost:8087
    description: Local development
...
paths:
...
  /api/content/countries:
    get:
      operationId: listCountries
      tags: [Content]
      summary: List available countries
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Available countries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_CountryList'
              example:
                success: true
                data:
                  - code: "NL"
                    name: "Nederland"
                    flagImage: "https://cdn.passtheo.nl/flags/nl.svg"
                    supportedLocales: ["nl", "en", "ar", "tr", "pl"]
                timestamp: "2026-03-19T10:00:00Z"
  /api/content/{countryCode}/product-types:
    get:
      operationId: listProductTypes
      tags: [Content]
      summary: List product types for a country (CBR, VCA, etc.)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CountryCode'
        - $ref: '#/components/parameters/TenantHeader'
      responses:
        '200':
          description: Product types
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ProductTypeList'
              example:
                success: true
                data:
                  - code: "cbr"
                    name: "CBR Rijexamen"
                    description: "Dutch driving theory exams"
                    icon: "https://cdn.passtheo.nl/icons/cbr.svg"
                    productCount: 3
                timestamp: "2026-03-19T10:00:00Z"
  /api/content/{countryCode}/{productTypeCode}/products:
    get:
      operationId: listProducts
      tags: [Content]
      summary: List products within a product type (Auto B, Motor A, etc.)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CountryCode'
        - $ref: '#/components/parameters/ProductTypeCode'
        - $ref: '#/components/parameters/TenantHeader'
      responses:
        '200':
          description: Products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ProductList'
              example:
                success: true
                data:
                  - code: "auto-b"
                    name: "Auto Theorie (B)"
                    licenceCode: "B"
                    description: "Car driving theory exam"
                    icon: "https://cdn.passtheo.nl/icons/auto-b.svg"
                    examConfig:
                      totalQuestions: 50
                      timeLimitMinutes: 30
                      passScore: 44
                    domainCount: 6
                    totalQuestions: 487
                timestamp: "2026-03-19T10:00:00Z"
  /api/content/{countryCode}/{productTypeCode}/{productCode}/domains:
    get:
      operationId: listDomains
      tags: [Content]
      summary: List domains for a product with user progress
      description: |
        Returns all domains with the current user's progress overlay
        (coverage %, accuracy %, strength label, locked/unlocked status).
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CountryCode'
        - $ref: '#/components/parameters/ProductTypeCode'
        - $ref: '#/components/parameters/ProductCode'
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/UserHeader'
        - $ref: '#/components/parameters/LocaleQuery'
      responses:
        '200':
          description: Domains with progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_DomainList'
              example:
                success: true
                data:
                  - code: "verkeersborden"
                    name: "Verkeersborden en verkeerstekens"
                    icon: "https://cdn.passtheo.nl/icons/verkeersborden.svg"
                    color: "#3B82F6"
                    topicCount: 7
                    questionCount: 85
                    isFreePreview: true
                    isLocked: false
                    progress:
                      coveragePercent: 72.0
                      accuracyPercent: 81.5
                      masteredCount: 42
                      strength: "STRONG"
                  - code: "voorrang"
                    name: "Voorrang en kruisingen"
                    icon: "https://cdn.passtheo.nl/icons/voorrang.svg"
                    color: "#EF4444"
                    topicCount: 6
                    questionCount: 78
                    isFreePreview: true
                    isLocked: false
                    progress:
                      coveragePercent: 45.0
                      accuracyPercent: 58.3
                      masteredCount: 12
                      strength: "WEAK"
                timestamp: "2026-03-19T10:00:00Z"
  /api/content/{countryCode}/{productTypeCode}/{productCode}/domains/{domainCode}/topics:
    get:
      operationId: listTopics
      tags: [Content]
      summary: List topics within a domain with user progress
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CountryCode'
        - $ref: '#/components/parameters/ProductTypeCode'
        - $ref: '#/components/parameters/ProductCode'
        - $ref: '#/components/parameters/DomainCode'
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/UserHeader'
        - $ref: '#/components/parameters/LocaleQuery'
      responses:
        '200':
          description: Topics with progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TopicList'
  /api/content/{countryCode}/{productTypeCode}/{productCode}/road-signs:
    get:
      operationId: listRoadSigns
      tags: [Content]
      summary: Road sign reference library
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CountryCode'
        - $ref: '#/components/parameters/ProductTypeCode'
        - $ref: '#/components/parameters/ProductCode'
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/LocaleQuery'
        - name: category
          in: query
          schema:
            type: string
            enum: [warning, prohibition, mandatory, information, priority, direction]
      responses:
        '200':
          description: Road signs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_RoadSignList'
  /api/content/{countryCode}/{productTypeCode}/{productCode}/lessons/{topicCode}:
    get:
      operationId: listLessons
      tags: [Content]
      summary: Theory lessons for a topic
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CountryCode'
        - $ref: '#/components/parameters/ProductTypeCode'
        - $ref: '#/components/parameters/ProductCode'
        - name: topicCode
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/LocaleQuery'
      responses:
        '200':
          description: Lessons
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_LessonList'

  # ─── PRACTICE SESSIONS ───
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantHeader:
      name: X-Tenant-ID
      in: header
      required: true
      schema:
        type: string
        format: uuid
    UserHeader:
      name: X-Keycloak-User-ID
      in: header
      required: true
      schema:
        type: string
        format: uuid
    CountryCode:
      name: countryCode
      in: path
      required: true
      schema:
        type: string
      example: "NL"
    ProductTypeCode:
      name: productTypeCode
      in: path
      required: true
      schema:
        type: string
      example: "cbr"
    ProductCode:
      name: productCode
      in: path
      required: true
      schema:
        type: string
      example: "auto-b"
    DomainCode:
      name: domainCode
      in: path
      required: true
      schema:
        type: string
      example: "verkeersborden"
    LocaleQuery:
      name: locale
      in: query
      schema:
        type: string
        default: "nl"
      description: Content locale (nl, en, ar, tr, pl)
  schemas:
...
    ApiResponse_CountryList:
      type: object
      properties:
        success: { type: boolean }
        data: { type: array, items: { $ref: '#/components/schemas/CountryDto' } }
        timestamp: { type: string, format: date-time }
    ApiResponse_ProductTypeList:
      type: object
      properties:
        success: { type: boolean }
        data: { type: array, items: { $ref: '#/components/schemas/ProductTypeDto' } }
        timestamp: { type: string, format: date-time }
    ApiResponse_ProductList:
      type: object
      properties:
        success: { type: boolean }
        data: { type: array, items: { $ref: '#/components/schemas/ProductDto' } }
        timestamp: { type: string, format: date-time }
    ApiResponse_DomainList:
      type: object
      properties:
        success: { type: boolean }
        data: { type: array, items: { $ref: '#/components/schemas/DomainWithProgressDto' } }
        timestamp: { type: string, format: date-time }
    ApiResponse_TopicList:
      type: object
      properties:
        success: { type: boolean }
        data: { type: array, items: { $ref: '#/components/schemas/TopicWithProgressDto' } }
        timestamp: { type: string, format: date-time }
    ApiResponse_RoadSignList:
      type: object
      properties:
        success: { type: boolean }
        data: { type: array, items: { $ref: '#/components/schemas/RoadSignDto' } }
        timestamp: { type: string, format: date-time }
    ApiResponse_LessonList:
      type: object
      properties:
        success: { type: boolean }
        data: { type: array, items: { $ref: '#/components/schemas/LessonDto' } }
        timestamp: { type: string, format: date-time }
...
    CountryDto:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        flagImage: { type: string }
        supportedLocales: { type: array, items: { type: string } }
    ProductTypeDto:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        description: { type: string }
        icon: { type: string }
        productCount: { type: integer }
    ProductDto:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        licenceCode: { type: string }
        description: { type: string }
        icon: { type: string }
        examConfig:
          type: object
          properties:
            totalQuestions: { type: integer }
            timeLimitMinutes: { type: integer }
            passScore: { type: integer }
        domainCount: { type: integer }
        totalQuestions: { type: integer }
    DomainWithProgressDto:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        icon: { type: string }
        color: { type: string }
        topicCount: { type: integer }
        questionCount: { type: integer }
        isFreePreview: { type: boolean }
        isLocked: { type: boolean }
        progress:
          type: object
          nullable: true
          properties:
            coveragePercent: { type: number }
            accuracyPercent: { type: number }
            masteredCount: { type: integer }
            strength: { type: string, enum: [UNKNOWN, WEAK, MODERATE, STRONG, MASTERED] }
    TopicWithProgressDto:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        difficulty: { type: string }
        questionCount: { type: integer }
        progress:
          type: object
          nullable: true
          properties:
            coveragePercent: { type: number }
            accuracyPercent: { type: number }
            masteredCount: { type: integer }
    RoadSignDto:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        signCategory: { type: string }
        description: { type: string }
        imageUrl: { type: string }
        shape: { type: string }
    LessonDto:
      type: object
      properties:
        title: { type: string }
        slug: { type: string }
        sections:
          type: array
          items:
            type: object
            properties:
              heading: { type: string, nullable: true }
              body: { type: string, nullable: true }
              tip: { type: string, nullable: true }
              keyRule: { type: string, nullable: true }
              relatedRoadSignCode: { type: string, nullable: true }
              sortOrder: { type: integer }
        summary: { type: string }
        coverImage: { type: string }
        videoUrl: { type: string }
        readTimeMinutes: { type: integer }
...